refactor(blockbinary): split the text layer out of the core (#1334 phase N)
The first header in this epic to actually hit both acceptance criteria.

    blockbinary.hpp alone   50832 -> 33280 lines   (-35%)
                               240 -> 174 headers
                        I/O-family:  5 -> 0

Under the 45000-line target with no iostream, sstream, iomanip, ostream or
istream in its graph. blockbinary.hpp keeps the arithmetic plus <string> for its
std::string constructor; manipulators.hpp takes type_tag/to_binary/to_hex/
to_decimal; iostream.hpp takes operator<<. Same naming as the posit pilot.

REMOVED THE friend operator<< DECLARATION, which is the durable part of this.
operator<< only calls to_decimal(), a public free function, so friendship was
never needed -- and the declaration was what let a translation unit missing the
include COMPILE and then fail to link. Without it the same mistake is a compile
error naming the missing function.

That mattered immediately: with the friend declaration, 2 TUs surfaced as link
errors in CI_LITE. Without it, 10 fail visibly at compile time. CI_LITE builds a
subset, so 8 would have reached the full tier undetected. Worth carrying into
blocktriple and every type in the rollout.

11 headers and 12 TUs now ask for text when they want text. <iosfwd> is gone from
blockbinary.hpp too -- with the friend declaration removed, no stream type
appears in the file at all.

METHOD, RECORDED BECAUSE IT COST FIVE BUILD CYCLES. I tried four times to find
the consumers by pattern-matching source: grep the type name, grep .bits(), grep
"<<" followed by .bits()/_block. Each missed cases, and the last miss shows why it
cannot work -- cfloat/table.hpp streams `signedValue`, a blockbinary variable, and
nothing on that line says so. A pattern over source text cannot know a variable's
type; only the compiler can.

The bad detour is worth recording too: one of those attempts shotgunned the
include into 35 headers, which would have pushed <sstream> and <iomanip> INTO
dbns_impl, e8m0_impl, efloat_impl and hfloat_impl -- four type cores -- inside a
PR whose purpose is removing them. Everything still compiled, which is exactly why
it was dangerous. Stripped back to the 11 headers where the need is proven.

For the rollout: build first, let the errors name the files, add includes only
where named. Grep is for estimating scope, never for proving completeness.

Verified: the 3 new headers and all 38 umbrellas compile on gcc 13.3, clang 18.1
AND riscv64-linux-gnu cross; every new header compiles standalone (one of them did
not, and nothing caught it because nothing included it); CI_LITE builds and passes
456/456; check-ascii clean; zero added lines over 120 columns.

Relates to #1334
