+++ include/sw/universal/number/integer/integer_impl.hpp
@@ -4,6 +4,7 @@
 // Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <universal/internal/blockbinary/manipulators.hpp>   // to_binary/to_hex on blockbinary (#1334)
 #include <string>
 #include <sstream>
 #include <iostream>
