+++ include/sw/universal/verification/blockbinary_test_status.hpp
@@ -3,6 +3,7 @@
 // Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <universal/internal/blockbinary/manipulators.hpp>   // to_binary/to_hex on blockbinary (#1334)
 #include <iostream>
 #include <iomanip>
 #include <universal/native/integers.hpp> // for to_binary(int)
