+++ internal/blockbinary/api/constexpr.cpp
@@ -10,6 +10,7 @@
 
 #include <universal/internal/blockbinary/blockbinary.hpp>
 #include <universal/verification/test_suite.hpp>
+#include <universal/internal/blockbinary/iostream.hpp>   // operator<< on blockbinary (#1334)
 
 int main()
 try {
